+++ web-components/src/components/donation-banner/donation-banner.ts
@@ -1,8 +1,7 @@
 import { LitElement, html, css } from "lit"
 import { customElement, property } from "lit/decorators.js"
 import { localized, msg, str } from "@lit/localize"
-import { getLocale } from "../../localization"
-import { getImageUrl } from "../../signals/app"
+import { getImageUrl, languageCode } from "../../signals/app"
 
 /**
  * Donation banner
@@ -48,7 +47,7 @@ export class DonationBanner extends LitElement {
   }
 
   get donateLink() {
-    const locale = getLocale()
+    const locale = languageCode.get()
     const link =
       locale in this.links ? this.links[locale as keyof typeof this.links] : this.links.default
     return this.getLinkWithQueryParams(link)
@@ -65,6 +64,7 @@ export class DonationBanner extends LitElement {
       border: 10px solid #ff6e78;
       width: 100%;
       background-color: white;
+      box-sizing: border-box;
     }
     @media (max-width: 732px) {
       .donation-banner,
@@ -146,6 +146,9 @@ export class DonationBanner extends LitElement {
       margin: 0.2rem 0rem 0.5rem;
       font-weight: 500;
     }
+    .donation-banner__main-div-wrapper {
+      color: black;
+    }
     @media (max-width: 732px) {
       .donation-banner__main-div-wrapper {
         margin-top: -25px;
@@ -163,7 +166,6 @@ export class DonationBanner extends LitElement {
     .donation-banner-footer__actions-section {
       display: flex;
       justify-content: space-around;
-      width: 100%;
       gap: 15px;
       padding: 0 10px;
     }