+++ web-components/src/components/mobile-badges/mobile-badges.ts
@@ -1,8 +1,7 @@
 import { LitElement, html, css, nothing } from "lit"
 import { customElement, state, property } from "lit/decorators.js"
 import { localized, msg } from "@lit/localize"
-import { getLocale } from "../../localization"
-import { getImageUrl } from "../../signals/app"
+import { getImageUrl, languageCode } from "../../signals/app"
 
 export type Badge = {
   href: string
@@ -260,9 +259,9 @@ export class MobileBadges extends LitElement {
    */
   getIosAppIconPath(language: string): string {
     if (language === "en") {
-      return "/appstore/black/appstore_UK.svg"
+      return "appstore/black/appstore_UK.svg"
     }
-    return `/appstore/black/appstore_${language.toLocaleUpperCase()}.svg`
+    return `appstore/black/appstore_${language.toLocaleUpperCase()}.svg`
   }
 
   /**
@@ -312,7 +311,7 @@ export class MobileBadges extends LitElement {
    * @returns The filtered list of badges.
    **/
   getFilteredBadges(): Badge[] {
-    const language = getLocale()
+    const language = languageCode.get()
     const badges: Badge[] = [
       {
         href: this.getAndroidAppLink(language),